copy util to index
